Debug Enemy
Added a debug enemy type to test enemy spawns. They are given the test sprite the player also uses right now.

The maze is now generated at runtime so that after each room is generated a random number of debug enemy spawners is attached to that room.

Movement and combat has yet to be added.